feat: better version resolution, logs, and docs
This commit makes a few quality of life changes.

First is correctly resolving the latest package
version from NPM when added with "titan add".

Second is an update to logging levels for
messages written to stderr in running tasks.
This applies to any command that runs scripts,
so they'll all benefit from more clarity knowing
this is not an error from Titan, but rather,
typically, a warning from the child process.

Lastly, some help text has been updated to
reflect the current usage and API for the CLI.